+++ src/utils/deleteItem.js
@@ -1,8 +1,8 @@
-import React from "react";
 import { without } from "./delArrItem";
+import { getTask } from "./getTask";
 
 export function deleteItem(todos) {
-  let tasks = JSON.parse(localStorage.getItem("tasks")) || [];
+  let tasks = getTask();
   tasks = without(tasks, todos);
   localStorage.setItem("tasks", JSON.stringify(tasks));
 }
